+++ sdk/src/applets/applet-scope.ts
