+++ sdk/src/shim.ts
@@ -1,25 +1,14 @@
 import { applets } from './index';
-import { type AppletFactory } from './main/applet-factory';
 import './index';
-import {
-  AppletReadyEvent,
-  AppletActionsEvent,
-  AppletDataEvent,
-  AppletResizeEvent,
-} from './events';
+import { type AppletFactory } from './applets/index';
+import { AppletEvent } from './events';
 
 declare global {
   interface Window {
     applets: AppletFactory;
-    AppletReadyEvent: typeof AppletReadyEvent;
-    AppletActionsEvent: typeof AppletActionsEvent;
-    AppletDataEvent: typeof AppletDataEvent;
-    AppletResizeEvent: typeof AppletResizeEvent;
+    AppletEvent: typeof AppletEvent;
   }
 }
 
 window.applets = applets;
-window.AppletReadyEvent = AppletReadyEvent;
-window.AppletActionsEvent = AppletActionsEvent;
-window.AppletDataEvent = AppletDataEvent;
-window.AppletResizeEvent = AppletResizeEvent;
+window.AppletEvent = AppletEvent;
